chore: remove more usages of contractkit, in debug screen (#5783)
### Description

Removing some unused functions and files, and also remove the latest
block item in the debug screen because it doesn't seem helpful anymore.
If we really wanted to know the latest block, we should print it for all
supported networks instead of Celo anyway.

### Test plan

n/a

### Related issues

- Related to RET-1173

### Backwards compatibility


Y

### Network scalability

If a new NetworkId and/or Network are added in the future, the changes
in this PR will:

- [ ] Continue to work without code changes, OR trigger a compilation
error (guaranteeing we find it when a new network is added)