chore: remove more usages of hardcoded colours (#5773)
### Description

I didn't do a thorough job of replacing hard coded colours in #5709 and
this PR is a follow up.

I removed the accent colors from the colors enum and intend to move the
remaining images that use custom hardcoded colors to the images
directory in a follow up.


### Test plan

n/a

### Related issues

- Related to RET-1155

### Backwards compatibility

Y

### Network scalability

If a new NetworkId and/or Network are added in the future, the changes
in this PR will:

- [ ] Continue to work without code changes, OR trigger a compilation
error (guaranteeing we find it when a new network is added)