feat(jumpstart): show add assets flow when no available Celo tokens (#5806)
### Description

This PR implements [this
design](https://www.figma.com/design/6sJ4fCxCptTybHB6ZZ7Pqi/Escrow-Feature?node-id=3324-2382&t=wNp4q9sJM0JyZxXQ-0).
- the live links flow should always work. so at the entry point, launch
the intro / add assets screen if the user does not have the correct
assets to continue.
- on the points and send screens, allow the live links flow to be
displayed even if the user has no tokens on Celo.
- improve the styles on the QR screen (the bottom components were too
close to the safe area before, and the address spilled awkwardly to 3
lines)

Not included in this PR:
- analytics
- network pre-selection on the swap and buy flows

### Test plan



https://github.com/user-attachments/assets/98dfac4f-0f6e-42bf-b228-395e9525b57b



https://github.com/user-attachments/assets/ce0ecf90-45b0-424d-82ae-fffc2533a3cf



### Related issues

- Related to RET-1177

### Backwards compatibility

Y

### Network scalability

If a new NetworkId and/or Network are added in the future, the changes
in this PR will:

- [ ] Continue to work without code changes, OR trigger a compilation
error (guaranteeing we find it when a new network is added)