refactor: create viem wallet without the need to unlock the keychain first (#5823)
### Description

This changes the viem wallet client creation so it can be called without
unlocking the keychain first.
Similar to how ContractKit is behaving. 

Unlocking is now necessary only when invoking an action that involves
signing from the account.

This means we can now call any wallet action without manually listing it
as a client extension.

Additionally, this tightens the security a bit more around the access to
the private key.
With the previous implementation it was possible to bypass unlocking
(after the first unlock) by accessing `client.account` directly.

### Test plan

- Updated tests

### Related issues

- Related to RET-1182

### Backwards compatibility

Yes

### Network scalability

If a new NetworkId and/or Network are added in the future, the changes
in this PR will:

- [x] Continue to work without code changes, OR trigger a compilation
error (guaranteeing we find it when a new network is added)