refactor(earn): use earn position in deposit screens (#5830)
### Description

Modifies the earn enter amount and bottom sheets to accept an earn
position object instead of tokenId.
Drops the info icon on the enter amount screen for multiple pools flow
per this
[conversation](https://valora-app.slack.com/archives/C03D3K379FE/p1724346415424529?thread_ts=1724283869.600569&cid=C03D3K379FE)
and adds provider logo.
Using the earn position's shortcut to deposit will be in a follow up

### Test plan

Manually by going through the earn flow with SHOW_MULTIPLE_POOLS gate
off and on.

### Related issues

- Part of ACT-1342

### Backwards compatibility

Yes

### Network scalability

If a new NetworkId and/or Network are added in the future, the changes
in this PR will:

- [x] Continue to work without code changes, OR trigger a compilation
error (guaranteeing we find it when a new network is added)