feat(dev): add screen that shows all image assets dynamically (#5825)
### Description

As the title.

Knip also complained about some unused images because the image path
changed, so i removed those.

### Test plan


https://github.com/user-attachments/assets/c1c3fff8-6d42-433a-9032-4696e92bf9d4



### Related issues

n/a

### Backwards compatibility

Y

### Network scalability

If a new NetworkId and/or Network are added in the future, the changes
in this PR will:

- [ ] Continue to work without code changes, OR trigger a compilation
error (guaranteeing we find it when a new network is added)