feat(settings): Update Initial settings screen (#5796)
### Description

For
[RET-1180](https://linear.app/valora/issue/RET-1180/[settings]-create-new-settings-screen).
First of several PRs for the Settings screen redesign. This PR largely
does the following:

* Adds the beginning of a new `SettingsMenu`, which will eventually
replace the current `ProfileMenu`. For now, this `SettingsMenu` is
inaccessible, and `ProfileMenu` is still used.
* Implements the required redesigns to a subset of the settings
submenus/screens, namely the Help and Wallet Address screens.

This PR only makes minor visible changes to the app today. These are:
* Some small design tweaks to the Help submenu and Contact Support
screen.
* Some small design tweaks to the QR code screen.
* Self-avatar style is now grey, as called for by designs, and a visual
bug with the text icon being off-center has been fixed.
* The margins on the horizontal rules between settings menu items are
now equal, rather than extending all the way to the right edge of the
screen.

### Test plan

Unit and manual tested. See video:



https://github.com/user-attachments/assets/0f55cb9e-71c1-4334-9be9-39e38063a8f9



### Related issues

Part of
[RET-1180](https://linear.app/valora/issue/RET-1180/[settings]-create-new-settings-screen).

### Backwards compatibility

Yes.

### Network scalability

If a new NetworkId and/or Network are added in the future, the changes
in this PR will:

- [x] Continue to work without code changes, OR trigger a compilation
error (guaranteeing we find it when a new network is added)