refactor: use viem in fiatconnect out review flow (#5835)
### Description

This PR makes the following changes to use viem instead of contractkit
in the FC transfer out flow:
- in the ReviewScreen, prepare a dummy send transfer to the user's own
wallet address since we don't yet know the provider's address. use the
prepared transactions for displaying the network fee / loading and error
states. pass the dummy serialised transaction to the sagas, to keep the
behaviour where we only invoke the `transfer` function on the provider
if the user confirms the transaction.
- in the sagas, use the fetched provider address to modify the tx data
and send the tx to the network via existing sagas that use viem.

Not included in this PR:
- updating the max amount allowed in the FC flow.

### Test plan

Update unit tests, but E2E tests should pass without tweaks.

### Related issues

- Related to RET-1183

### Backwards compatibility

Y

### Network scalability

If a new NetworkId and/or Network are added in the future, the changes
in this PR will:

- [ ] Continue to work without code changes, OR trigger a compilation
error (guaranteeing we find it when a new network is added)