feat(settings): Implement new settings submenu designs (#5816)
### Description

For
[RET-1180](https://linear.app/valora/issue/RET-1180/[settings]-create-new-settings-screen).
This PR implements another chunk of the redesign to the settings menu.
In particular, this PR updates the designs/implements submenus for the
following menu options:

* Invite
* Connected Dapps
* Legal

Both the Invite and Connected Dapps screen currently exist and are
accessible in the app, so the changes to those screens will be reflected
in the app, despite the new settings screen not yet being connected.

### Test plan

Unit and manual tested. See video below:



https://github.com/user-attachments/assets/0cbd2a61-076e-4952-936d-971f0bf508b1


### Related issues

- Part of
[RET-1180](https://linear.app/valora/issue/RET-1180/[settings]-create-new-settings-screen).

### Backwards compatibility

Yes.

### Network scalability

If a new NetworkId and/or Network are added in the future, the changes
in this PR will:

- [x] Continue to work without code changes, OR trigger a compilation
error (guaranteeing we find it when a new network is added)