feat(earn): Add info bottom sheets to pool detail screen (#5854)
### Description

Bottom sheets added for TVL, age and Yield Rate. Follow up PR will add
bottom sheet for deposit & earnings.

Updated [this
ticket](https://linear.app/valora/issue/ACT-1306/update-earnhome-to-use-custom-header-to-avoid-white-bar-on-top-of)
for header fix because same issue as EarnHome.

### Test plan

Manual:


https://github.com/user-attachments/assets/c2490e96-4865-42bb-bc70-65d953bce36e



https://github.com/user-attachments/assets/62906582-8c63-4d10-acda-734f54a94b72



### Related issues

- Part of ACT-1323

### Backwards compatibility

Yes

### Network scalability

If a new NetworkId and/or Network are added in the future, the changes
in this PR will:

- [X] Continue to work without code changes, OR trigger a compilation
error (guaranteeing we find it when a new network is added)