refactor: copy account module from @celo/cryptographic-utils@3.2.0 (#5861)
### Description

This is the first step to remove `@celo/cryptographic-utils`.

It copies the
[accounts](https://github.com/celo-org/celo-monorepo/blob/sdks-3.2.0/packages/sdk/cryptographic-utils/src/account.ts)
module from it.

I've copied the exact same version used from the celo SDK: `3.2.0`.
And listed the exact same NPM packages it was using internally.
This is to ensure it behaves the same.
We'll upgrade/remove these in a follow up PR.

I've slightly modified it to directly include some sub modules, which we
didn't need to share.
- `@celo/base/lib/account` (some type definitions)
- `@celo/base/lib/string` (only 1 function used)

### Test plan

- Tests pass

### Related issues

- Part of RET-1185

### Backwards compatibility

Yes

### Network scalability

If a new NetworkId and/or Network are added in the future, the changes
in this PR will:

- [x] Continue to work without code changes, OR trigger a compilation
error (guaranteeing we find it when a new network is added)