feat(settings): Hook up new settings screen, remove old one (#5840)
### Description

Completes
[RET-1180](https://linear.app/valora/issue/RET-1180/%5Bsettings%5D-create-new-settings-screen).
Hooks up the new settings screen to the rest of the app, and removes the
old one from the codebase entirely.

### Test plan

Unit and manual tested. See video below:



https://github.com/user-attachments/assets/ec5e53ef-7d4a-42d0-9227-154f3432da8b


### Related issues

- Fixes
[RET-1180](https://linear.app/valora/issue/RET-1180/%5Bsettings%5D-create-new-settings-screen).

### Backwards compatibility

Yes.

### Network scalability

If a new NetworkId and/or Network are added in the future, the changes
in this PR will:

- [x] Continue to work without code changes, OR trigger a compilation
error (guaranteeing we find it when a new network is added)