feat(onbaording): Parameterize onboarding features via config (#5859)
### Description

For
[RET-1152](https://linear.app/valora/issue/RET-1152/[onboarding]-simplify-onboarding-steps-make-configurable).
Updates onboarding to parameterize certain optional features via local
config. I say "features" here instead of "screens", since this includes
which kinds of backup/restore are offered, which doesn't cleanly map to
a single screen.

The configuration for most of the onboarding features is currently
handled via Statsig, however, for new app versions, the values for these
features are fixed, and I don't anticipate us needing to change them in
the near future, so I've taken this as an opportunity to replace them
with local config. Having _both_ local config and config from Statsig
simaltaneously doesn't really work, since in all cases we'd just have to
defer to one (likely Statsig) over the other.

If we want to keep Statsig as the config source for these features, I
think our best bet would be to wait to do this chunk of work until after
the fork, since there's little value in us doing it here if we can never
actually use the config in Valora, since we'd always prefer to use
Statsig instead. Otherwise, I think some approach like the one offered
in this PR probably makes sense.

### Test plan

Unit and manual tested.

### Related issues

- Fixes
[RET-1152](https://linear.app/valora/issue/RET-1152/[onboarding]-simplify-onboarding-steps-make-configurable).

### Backwards compatibility

Yes.

### Network scalability

If a new NetworkId and/or Network are added in the future, the changes
in this PR will:

- [x] Continue to work without code changes, OR trigger a compilation
error (guaranteeing we find it when a new network is added)