refactor: remove crypto-js which is now deprecated (#5866)
### Description

This removes [crypto-js](https://github.com/brix/crypto-js) which is now
discontinued:
> Active development of CryptoJS has been discontinued. This library is
no longer maintained.
>
> Nowadays, NodeJS and modern browsers have a native Crypto module. The
latest version of CryptoJS already uses the native Crypto module for
random number generation, since Math.random() is not crypto-safe.
Further development of CryptoJS would result in it only being a wrapper
of native Crypto. Therefore, development and maintenance has been
discontinued, it is time to go for the native crypto module.

Bonus, because it now uses the native crypto module (via
[react-native-quick-crypto](https://github.com/margelo/react-native-quick-crypto))
AES encryption/decryption should be faster. Though the gains are
probably not noticeable given the small use we have.

Note: there's an important change for tests too, they now all use real
encryption.
- I feel this is less surprising than what we used to do.
- I was able to refactor tests that expected mocked encryption without
too much trouble

### Test plan

- Updated tests (ensuring backward compat with CryptoJS encrypted
strings)
- Manually tested existing keychain items encrypted by CryptoJS can
still be decrypted in Valora

### Related issues

- Part of RET-1185

### Backwards compatibility

Yes

### Network scalability

If a new NetworkId and/or Network are added in the future, the changes
in this PR will:

- [x] Continue to work without code changes, OR trigger a compilation
error (guaranteeing we find it when a new network is added)