chore: remove unused code related to ContractKit (#5897)
### Description

This removes some unused code related to ContractKit.

There will be more to come related to removing `KeychainWallet`, but
this focuses on parts which were unused but not dependent on it, to make
it easier to review.

### Test plan

- Tests pass

### Related issues

- Part of RET-1192

### Backwards compatibility

Yes

### Network scalability

If a new NetworkId and/or Network are added in the future, the changes
in this PR will:

- [x] Continue to work without code changes, OR trigger a compilation
error (guaranteeing we find it when a new network is added)

---------

Co-authored-by: Joseph Bergeron <jophish126@gmail.com>