chore: remove `@celo/contractkit`, `web3` and related unused code (#5904)
### Description

It's finally happening! 🥲 🎉 
You served us well `@celo/contractkit`, but now it's time to go.

Also the old `web3` can go too!

Big shout out to the whole Valora eng team who made this possible 🙌 

### Test plan

- Tests pass

### Related issues

- Part of RET-1192

### Backwards compatibility

Yes

### Network scalability

If a new NetworkId and/or Network are added in the future, the changes
in this PR will:

- [x] Continue to work without code changes, OR trigger a compilation
error (guaranteeing we find it when a new network is added)

---------

Co-authored-by: Joseph Bergeron <jophish126@gmail.com>