feat(config): Parameterize whether or not fiatconnect quotes are supported (#5919)
### Description

Adds a static config parameter to enable/disable fetching FiatConnect
quotes.

### Test plan

<!-- Demonstrate the change is solid, or why it doesn't need testing.
Example: add any manual testing steps or scenarios (if not obvious),
screenshots / videos if the pull request changes the user interface.
-->

### Related issues

- Fixes
[RET-1145](https://linear.app/valora/issue/RET-1145/[fiatconnect]-make-fetching-fc-quotes-configurable-in-valora)

### Backwards compatibility

Yes

### Network scalability

If a new NetworkId and/or Network are added in the future, the changes
in this PR will:

- [x] Continue to work without code changes, OR trigger a compilation
error (guaranteeing we find it when a new network is added)