fix(earn): Update providerId for cached earn txs (#5867)
### Description

Updated providerId for earn transactions from `aave-v3` to `aave` in the
backend, but it didn't update the cached transactions. This should
update the providerId for cached earn transactions stored in root state.

### Test plan

Ran app locally and confirmed errors went away, pool provider name
showed correctly

### Related issues

- Fixes N/A

### Backwards compatibility

Yes

### Network scalability

If a new NetworkId and/or Network are added in the future, the changes
in this PR will:

- [X] Continue to work without code changes, OR trigger a compilation
error (guaranteeing we find it when a new network is added)