fix(onbaording): Disable CPV in onboarding (#5931)
### Description

We previously ported onboarding params to be locally configured as
preparation for MS development, but forgot to set the environment
variables as to disable CPV in onboarding.

### Test plan

Manual tested.

### Related issues

### Backwards compatibility

Y

### Network scalability

If a new NetworkId and/or Network are added in the future, the changes
in this PR will:

- [x] Continue to work without code changes, OR trigger a compilation
error (guaranteeing we find it when a new network is added)