ci: ignore emergetools errors (#5944)
### Description

The new Emergetools indie plan is limited in the number of builds.
Ignore errors for now until we decide what we want to do with it.

### Test plan

- Tests pass

### Related issues

N/A

### Backwards compatibility

Yes

### Network scalability

If a new NetworkId and/or Network are added in the future, the changes
in this PR will:

- [x] Continue to work without code changes, OR trigger a compilation
error (guaranteeing we find it when a new network is added)