fix: restore Cloud Account Backup (#5941)
### Description

Cloud Account Backup broke in
https://github.com/valora-inc/wallet/pull/5866.

That PR changed `global.crypto` to be polyfilled by
`react-native-quick-crypto` and the version of `@toruslabs/eccrypto`
pulled by our Cloud Account Backup, wasn't working out-of-the box with
it.

However it got fixed in
https://github.com/torusresearch/eccrypto/pull/40 and landed in
`@toruslabs/eccrypto >= 5.0.2`.

So this PR upgrades the `@toruslabs` libs so it pulls the latest version
of `@toruslabs/eccrypto@5.0.4` which works with
`react-native-quick-crypto`.

It's using new major versions. Scanning through the [release
notes](https://github.com/torusresearch/torus.js/releases), the main
impact for us was the way to import the module. But let me know if
there's an important detail I've missed.

Additionally, I found there was a bug in `react-native-quick-crypto`,
`getAuthTag` was returning an `ArrayBuffer` instead of a `Buffer`
leading to an incorrect encoding in the final encrypted string.
I contributed a fix upstream
(https://github.com/margelo/react-native-quick-crypto/pull/443) and
created a patch to use until it's merged. And made a slight change to
the wallet code so we can decrypt data if the `authTag` was serialized
as an `ArrayBuffer` (e.g. `"1,2,3,4,..."`) instead of the expected
base64.

This replaces https://github.com/valora-inc/wallet/pull/5934

### Test plan

- Tests pass
- Manually tested:
  - Restore from Cloud Account Backup works 
  - Setup Cloud Account Backup works
  - Delete Cloud Account Backup works

### Related issues

See Slack: [thread
1](https://valora-app.slack.com/archives/C04B61SJ6DS/p1725642273567639),
[thread
2](https://valora-app.slack.com/archives/C025X4WJT09/p1725666426992419).

### Backwards compatibility

Yes

### Network scalability

If a new NetworkId and/or Network are added in the future, the changes
in this PR will:

- [x] Continue to work without code changes, OR trigger a compilation
error (guaranteeing we find it when a new network is added)