chore: update `@th3rdwave/react-navigation-bottom-sheet` (#5930)
### Description

Update `@th3rdwave/react-navigation-bottom-sheet` to the latest version
and simplify some usage.
It was updated to work with the latest version of
`gorhom/react-native-bottom-sheet` with dynamic sizing.
See
https://github.com/th3rdwave/react-navigation-bottom-sheet/releases/tag/v0.3.0

I looked at the previous changes where we added workarounds for screen
usage:
- https://github.com/valora-inc/wallet/pull/5062
- https://github.com/valora-inc/wallet/pull/5485

These aren't needed anymore, but there's a very small patch needed
again. I'll contribute it upstream.
However, bonus is we now get scroll down to dismiss for bottom sheet
screens!

### Test plan

- Tests pass
- Manually checked bottom sheets display with the correct size on both
iOS and Android. And behave as expected (scroll to dismiss, etc)
  - Hopefully I haven't missed any special case 👀 

### Related issues

N/A

### Backwards compatibility

Yes

### Network scalability

If a new NetworkId and/or Network are added in the future, the changes
in this PR will:

- [x] Continue to work without code changes, OR trigger a compilation
error (guaranteeing we find it when a new network is added)

---------

Co-authored-by: renovate[bot] <29139614+renovate[bot]@users.noreply.github.com>