chore: migrate min required app version to Statsig (#5957)
### Description

As the title. There should be no functional difference, except the
statsig value is refreshed on app start / foregrounded whereas the
firebase value change was immediately updated. Seems like an unimpactful
detail.

### Test plan

Unit tests

### Related issues

- Fixes RET-1150

### Backwards compatibility

Y

### Network scalability

If a new NetworkId and/or Network are added in the future, the changes
in this PR will:

- [ ] Continue to work without code changes, OR trigger a compilation
error (guaranteeing we find it when a new network is added)