feat(ci): Setup CI, get e2e tests working (#1)
### Description

This PR gets the e2e tests working, and sets up the Namespace
integration for the Android runners. ATM, all enabled e2e tests pass
except for the WC tests, though this perhaps seems unrelated to this
repo in particular? [See Slack
thread](https://valora-app.slack.com/archives/C025V1D6F3J/p1726096062541629).

### Test plan

<!-- Demonstrate the change is solid, or why it doesn't need testing.
Example: add any manual testing steps or scenarios (if not obvious),
screenshots / videos if the pull request changes the user interface.
-->

### Related issues

- Fixes #[issue number here]

### Backwards compatibility

<!-- Brief explanation of why these changes are/are not backwards
compatible. -->

### Network scalability

If a new NetworkId and/or Network are added in the future, the changes
in this PR will:

- [ ] Continue to work without code changes, OR trigger a compilation
error (guaranteeing we find it when a new network is added)