refactor: move deep link prefix to .env file (#6008)
### Description

Related to making the MS runtime native folders brand agnostic. This PR:
- add deep link prefix value to .env files
- read the variable during build and run times
- update the config name in the `src` folder 

### Test plan

CI passes - we have deep link tests there.

### Related issues

- Related to RET-1157

### Backwards compatibility

Y

### Network scalability

If a new NetworkId and/or Network are added in the future, the changes
in this PR will:

- [ ] Continue to work without code changes, OR trigger a compilation
error (guaranteeing we find it when a new network is added)