refactor: add env variable for user agent app name (#6009)
### Description

Related to making the native build folders brand agnostic. This PR:
- add user agent app name in .env files. according to RET-558,
downstream services expect a certain value here so i didn't use the
APP_DISPLAY_NAME.
- update prettier config - i was getting errors with the pre-commit
hook, and it seems [the plugin search feature has been removed in v3 of
prettier](https://prettier.io/blog/2023/07/05/3.0.0.html#plugin-search-feature-has-been-removed-14759httpsgithubcomprettierprettierpull14759-by-fiskerhttpsgithubcomfisker)
so we need to declare them explicitly.

### Test plan

Tested the user agent header values on iOS and Android apps manually

### Related issues

- Related to RET-1157

### Backwards compatibility

Y

### Network scalability

If a new NetworkId and/or Network are added in the future, the changes
in this PR will:

- [ ] Continue to work without code changes, OR trigger a compilation
error (guaranteeing we find it when a new network is added)