feat(earn): show swappable tokens on enter amount screen (#5998)
### Description

Part 1 of implementing swap and deposit on the enter amount screen. Adds
token bottom sheet for selecting a swappable token. If only one token is
eligible, defaults to it and no dropdown is shown. Preparing swap +
deposit tx will be in a follow up PR

### Test plan

Unit tests, manual (by forcing swap-deposit flow)



https://github.com/user-attachments/assets/707fa7c2-10cb-484b-aad1-d0b9c566674d



### Related issues

- Part of ACT-1356

### Backwards compatibility

Yes

### Network scalability

If a new NetworkId and/or Network are added in the future, the changes
in this PR will:

- [ ] Continue to work without code changes, OR trigger a compilation
error (guaranteeing we find it when a new network is added)

---------

Co-authored-by: Tom McGuire <mcgtom10@gmail.com>