feat(earn): prepare swap + deposit transaction (#5999)
### Description

Part 2 of swap + deposit. Builds on #5998. Update existing
`prepareDepositTransactions` (renamed from `prepareSupplyTransactions`)
to prepare either `deposit` or `swap-deposit` transactions. Displaying
swap info will be in the next PR

### Test plan

CI, manual

### Related issues

- Part of ACT-1356

### Backwards compatibility

Yes

### Network scalability

If a new NetworkId and/or Network are added in the future, the changes
in this PR will:

- [x] Continue to work without code changes, OR trigger a compilation
error (guaranteeing we find it when a new network is added)