feat(send): Remove CPV from send flow (#12)
### Description

For
[RET-1146](https://linear.app/valora/issue/RET-1146/[shell]-update-sendreceive-flow).
Removes contact + CPV + secure send functionality from the send flow.
This is accomplished _not_ by gutting the wallet of anything related to
CPV, but instead by removing any possibility of accessing a feature
which interacts with the CPV system. This is a bit of a tradeoff, as it
leads to a decent chunk of unused code just kinda hanging around, but I
opted for this approach for the following reasons:
* Actually removing all the CPV-related code would make it very
difficult and cumbersome to re-introduce it if we ever wanted to add CPV
support to Mobile Stack, which is a real possibility.
* Removing all the code would introduce a huge diff between the MS
Runtime and the Valora wallet, which we'd really like to avoid if we
can. The CPV system touches a lot of parts of the app.
* This approach is just much easier than the alternative; gutting the
app of CPV code would have been a much greater task.

### Test plan

Unit tests updated (or skipped, where appropriate). See video below:


https://github.com/user-attachments/assets/94f24f05-a791-4392-8893-e820aaed1232

### Related issues

- Fixes
[RET-1146](https://linear.app/valora/issue/RET-1146/[shell]-update-sendreceive-flow).

### Backwards compatibility

Yes.

### Network scalability

If a new NetworkId and/or Network are added in the future, the changes
in this PR will:

- [x] Continue to work without code changes, OR trigger a compilation
error (guaranteeing we find it when a new network is added)