chore(TransactionFeedV2): Cleanup stand by transactions in Transaction Feed V2 (#6146)
### Description
4th PR of RET-1207. This implements the cleanup of confirmed stand by
transactions that are already present in the feed. Currently, we keep
confirmed transactions in `standByTransactions` even if they are present
in the pagination data. This is unnecessary as it takes extra space from
the persisted storage.

### Test plan
- changed the test that checks merging of pages with stand by
transactions
- added test to check if the stand by transaction was correctly removed
if it appeared in the pagination data

### Related issues

- Relates to RET-1207

### Backwards compatibility
Yes

### Network scalability

If a new NetworkId and/or Network are added in the future, the changes
in this PR will:

- [x] Continue to work without code changes, OR trigger a compilation
error (guaranteeing we find it when a new network is added)