chore(TransactionFeedV2): Show stand by transactions when paginated data is empty (#6152)
### Description
6th PR for RET-1207. Adds handling of the empty transaction feed. There
is a possibility that a new user can have a few pending/failed
transactions that are stored in `standByTransactions` but paginated data
returning empty array for the first page. In this case we need to show
stand by transactions.

### Test plan
Adds test to show stand by transactions if the paginated data is empty.

### Related issues

- Relates to RET-1207

### Backwards compatibility
Yes

### Network scalability

If a new NetworkId and/or Network are added in the future, the changes
in this PR will:

- [x] Continue to work without code changes, OR trigger a compilation
error (guaranteeing we find it when a new network is added)