chore(TransactionFeedV2): Add "No more transactions" toast  (#6153)
### Description
7th PR for RET-1207. Add "No more transactions" toast when:
- the first page is the only page and there's decent amount of
transactions to trigger the scroll
- next page returns an empty array

This behaviour will be changed in the follow-up PR once connected to the
new blockchain-api handler.

### Test plan
Adds tests to check the new toast behaviour.

### Related issues
- Relates to RET-1207

### Backwards compatibility
Yes

### Network scalability

If a new NetworkId and/or Network are added in the future, the changes
in this PR will:

- [x] Continue to work without code changes, OR trigger a compilation
error (guaranteeing we find it when a new network is added)