chore(TransactionFeedV2): Re-write transactions reducer to slice (#6166)
### Description
11th PR for RET-1207. Rewrite of transactions `reducer.ts/actions.ts`
files to `slice.ts/selectors.ts`. Slice also requires all actions to get
arguments passed as a single `payload` object so needed to refactor all
the usages of the corresponding actions. For better review experience,
each action usage refactor was merged as a separate commit.

This rewrite is necessary to move a bunch of redux-related `useEffects`
from TransactionFeedV2 component to `extraReducers`.

Also, this rewrite doesn't change the implementation of
TransactionFeedV2 as it is already a bit of a big PR.

### Test plan
Re-use existing reducer.test.ts and rename it to slice.test.ts

### Related issues

- Relates to RET-1207

### Backwards compatibility
Yes

### Network scalability

If a new NetworkId and/or Network are added in the future, the changes
in this PR will:

- [x] Continue to work without code changes, OR trigger a compilation
error (guaranteeing we find it when a new network is added)