chore(TransactionFeedV2): Merge old stand by transactions that are out of scope of pagination (#6185)
### Description
This PR fixes the merging process of pagination data with stand by
transactions. Previously, there were still some stand by transactions
that didn't get merged, specifically those that are older than the last
transaction from pagination. This PR fixes that by merging all the stand
by transactions that are "out of scope" of pagination at the very end of
the feed.

### Test plan
Added test that checks proper merging for old stand by transactions

### Related issues

- Relates to RET-1207

### Backwards compatibility
Yes

### Network scalability

If a new NetworkId and/or Network are added in the future, the changes
in this PR will:

- [x] Continue to work without code changes, OR trigger a compilation
error (guaranteeing we find it when a new network is added)