refactor: Use mobilestack references in android builds (#6039)
### Description

Makes the Android builds more brand agnostic. Replaces the existing
keystore android keystore containing the upload key by:
* Renaming the private key alias within the keystore from
`celo-release-key` to `mobilestack-release-key`
* Using the same password for both the store as well as the private key
within

### Test plan

<!-- Demonstrate the change is solid, or why it doesn't need testing.
Example: add any manual testing steps or scenarios (if not obvious),
screenshots / videos if the pull request changes the user interface.
-->

### Related issues

- Fixes #[issue number here]

### Backwards compatibility

<!-- Brief explanation of why these changes are/are not backwards
compatible. -->

### Network scalability

If a new NetworkId and/or Network are added in the future, the changes
in this PR will:

- [ ] Continue to work without code changes, OR trigger a compilation
error (guaranteeing we find it when a new network is added)