fix(TransactionFeedV2): Fix removals and "jumps" of pending transactions  (#6206)
### Description
This PR fixes an issue when some pending transactions can be removed
from the feed for a short time and re-added back again. This was
happening because pending and confirmed stand by transactions were split
into 2 states which were independently affecting different parts of the
final `sections` state.

Example:
- `pending` transaction appears with a spinning indicator after a Send
action
- when it is marked as confirmed it changes both
`standByTransactions.pending` and `standByTransactions.confirmed`

This causes the following update chains: 
- `standByTransactions.pending` -> `sections`
- `standByTransactions.confirmed` -> `updatePaginatedData` ->
`paginatedData` -> `confirmedTransactions` -> `sections`

So in result `sections` is updated twice with visual feedback in a form
of removing and re-adding back transactions.

### Test plan
All existing tests pass.


https://github.com/user-attachments/assets/1a34e993-9621-4f2d-80e0-d7fcc2069cdb




### Related issues

- Relates to RET-1207

### Backwards compatibility

<!-- Brief explanation of why these changes are/are not backwards
compatible. -->

### Network scalability

If a new NetworkId and/or Network are added in the future, the changes
in this PR will:

- [x] Continue to work without code changes, OR trigger a compilation
error (guaranteeing we find it when a new network is added)