chore(prices): fetch exchange rate from cloud function  (#6216)
### Description

Use cloud function `getExchangeRate` instead of the GraphQL endpoint of
the blockchain-api.

Not hiding it behind the feature flag since the risk of failure seems
low, and it can be handled by the backend update.

### Test plan

- Updated unit test

### Related issues

- Fixes RET-1224

### Backwards compatibility

Y

### Network scalability

If a new NetworkId and/or Network are added in the future, the changes
in this PR will:

- [x] Continue to work without code changes, OR trigger a compilation
error (guaranteeing we find it when a new network is added)