chore(balances): fetch token balances from cloud function (#6215)
### Description

Use cloud function `getWalletBalances` instead of the GraphQL endpoint
of the blockchain-api.

Not hiding it behind the feature flag since the risk of failure seems
low, and it can be handled by the backend update.

### Test plan

- Updated unit test

### Related issues

- Part of RET-1224

### Backwards compatibility

Y

### Network scalability

If a new NetworkId and/or Network are added in the future, the changes
in this PR will:

- [x] Continue to work without code changes, OR trigger a compilation
error (guaranteeing we find it when a new network is added)

---------

Co-authored-by: Jean Regisser <jean.regisser@gmail.com>