chore: use comma separated query params and only query enabled networks (#6219)
### Description

Updating some query params for `getWalletTransactions` and
`getWalletBalances`
Following the new accepted format done in
https://github.com/valora-inc/blockchain-api/pull/766 and
https://github.com/valora-inc/blockchain-api/pull/769.

It's slightly less verbose when looking at logs.

Also updated `getWalletTransactions` to only request networkIds enabled
in statsig.

### Test plan

- Updated tests
- Manually checked

### Related issues

- See also this Slack
[thread](https://valora-app.slack.com/archives/C029Z1QMD7B/p1731057942380629?thread_ts=1731057749.819289&cid=C029Z1QMD7B)

### Backwards compatibility

Yes

### Network scalability

If a new NetworkId and/or Network are added in the future, the changes
in this PR will:

- [x] Continue to work without code changes, OR trigger a compilation
error (guaranteeing we find it when a new network is added)